Refactored code structure of ControlsComponent
